WIP: Move GoogleOAuth class to own file
